+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade `golioth/golioth-zephyr-boards` dependency to [`v1.0.1`](https://github.com/golioth/golioth-zephyr-boards/tree/v1.0.1)
 - Dependencies use https instead of ssh GitHub URLs
 - libostentus removed from code base and included as a Zephyr module
+- renamed app_work.c/h to app_sensors.c/h
 
 ### Fixed
 - Fix build error when `CONFIG_LIB_OSTENTUS=n` on the `aludel_mini_v1_sparkfun9160` board.